"endpoint REST"

+++ order-service/src/main/java/br/com/microservices/orchestrated/orderservice/core/document/Order.java
@@ -23,5 +23,5 @@ public class Order {
     private LocalDateTime createdAt;
     private String transactionId;
     private double totalAmount;
-    private double totalItems;
+    private int totalItems;
 }
